Mirror the Attestation Framework predicates package structure
