+++ go.mod
@@ -4,7 +4,7 @@ go 1.20
 
 require (
 	github.com/google/go-cmp v0.6.0
-	github.com/in-toto/attestation v0.1.1-0.20230908124000-cefdc8c70608
+	github.com/in-toto/attestation v1.0.1
 	github.com/secure-systems-lab/go-securesystemslib v0.7.0
 	github.com/shibumi/go-pathspec v1.3.0
 	github.com/spf13/cobra v1.8.0
@@ -12,6 +12,7 @@ require (
 	github.com/stretchr/testify v1.8.4
 	golang.org/x/sys v0.15.0
 	google.golang.org/grpc v1.60.1
+	google.golang.org/protobuf v1.31.0
 )
 
 require (
@@ -32,6 +33,5 @@ require (
 	golang.org/x/text v0.14.0 // indirect
 	golang.org/x/tools v0.6.0 // indirect
 	google.golang.org/genproto/googleapis/rpc v0.0.0-20231002182017-d307bd883b97 // indirect
-	google.golang.org/protobuf v1.31.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
